requirements.txt
transformers==4.40.0
pydub==0.25.1
sounddevice==0.4.6
torch==2.3.0
numpy==1.26.4
wavio==0.0.8
keyboard
paho-mqtt
scipy